+++ lib/constants.js
@@ -1,4 +1,5 @@
 import toHash from 'tohash'
+import dynamic from 'next/dynamic'
 
 export const FONTS = [
   { id: 'Anonymous Pro', name: 'Anonymous Pro' },
@@ -908,18 +909,11 @@ const unfold = (f, seed) => {
 }`
 
 if (typeof window !== 'undefined' && typeof window.navigator !== 'undefined') {
-  const alreadyLoaded = new Set()
+  require(`codemirror/mode/javascript/javascript`)
 
-  LANGUAGES.filter(language => language.mode !== 'auto' && language.mode !== 'text').forEach(
-    language => {
-      if (language.mode && !alreadyLoaded.has(language.mode)) {
-        alreadyLoaded.add(language.mode)
-        language.custom
-          ? require(`./custom/modes/${language.mode}`)
-          : require(`codemirror/mode/${language.mode}/${language.mode}`)
-      }
-    }
-  )
+  const alreadyLoaded = new Set(['javascript'])
+
+  dynamic(() => import('./modes')(alreadyLoaded))
 }
 
 export const DEFAULT_SETTINGS = {
